refactor(popup-menu)!: rename `segment` to `pathSegment` and `resolve/` to `menu-tree/`
