+++ packages/react/src/context-menu/root/root.tsx
@@ -12,7 +12,7 @@ import {
   usePopupMenuRoot,
   type VirtualAnchor,
 } from '../../internal/popup-menu/index.js'
-import type { GetRowIdFn } from '../../internal/popup-menu/resolve/types.js'
+import type { GetRowIdFn } from '../../internal/popup-menu/menu-tree/types.js'
 import type {
   ContextMenuHighlightChangeEventDetails,
   ContextMenuOpenChangeEventDetails,
