+++ packages/react/src/internal/popup-menu/deep-search/data-subpages.tsx
@@ -6,8 +6,8 @@ import {
   defaultGetRowId,
   isPopupMenuNode,
   resolveDetachedNode,
-} from '../resolve/resolve.js'
-import type { PopupMenuNode } from '../resolve/types.js'
+} from '../menu-tree/resolve.js'
+import type { PopupMenuNode } from '../menu-tree/types.js'
 import { useAsyncMenuCoordinator } from './async-coordinator.js'
 import { useDataPopupContext } from './context.js'
 import { warnOutOfTreeDef } from './data-list.js'
